auto-improve: [#799 Step 6/6] Extract `cmd_cycle` and `cmd_dispatch` from `cai.py` → `cai_lib/cmd_cycle.py`; reduce `cai.py` to pure dispatcher

Refs damien-robotsix/robotsix-cai#823
